feat: prepare code to work with additional entries [YTFRONT-4122]
All changes in src/ui just break some dependencies to avoud including unnecessary code to a custom entry.
