--- packages/ui/src/ui/components/index.tsx
@@ -1,3 +1,0 @@
-export {default as ChartLink} from './ChartLink/ChartLink';
-export {default as Icon} from './Icon/Icon';
-export {default as Link} from './Link/Link';
