+++ packages/ui/src/ui/hooks/useAppSelector.ts
@@ -1,3 +1,3 @@
 import {TypedUseSelectorHook, useSelector} from 'react-redux';
-import {StoreType} from '../store';
+import {StoreType} from '../store/store.main';
 export const useAppSelector: TypedUseSelectorHook<StoreType> = useSelector;
