--- packages/ui/src/ui/legacy-styles/elements/base.scss
@@ -1,3 +1,0 @@
-*, *::before, *::after {
-    box-sizing: border-box;
-}
