+++ packages/ui/src/ui/legacy-styles/elements/elements.scss
@@ -1,4 +1,3 @@
-@import 'base';
 @import 'ellipsis';
 @import 'link';
 @import 'monospace';
