--- packages/ui/src/ui/lib.tsx
@@ -1,7 +1,0 @@
-export {renderApp} from './render-app';
-export type {UIFactory, UIFactoryClusterPageInfo, UIFactoryRootPageInfo} from './UIFactory';
-export {isDocsAllowed} from './config';
-export {default as YT} from './config/yt-config';
-
-export * from '../shared/lib';
-export * from './components';
