+++ packages/ui/src/ui/store/actions/global/experimental-pages.ts
@@ -1,0 +1,15 @@
+import _ from 'lodash';
+
+import {getCurrentUserName} from '../../../store/selectors/global';
+import {GLOBAL_PARTIAL} from '../../../constants/global';
+import UIFactory from '../../../UIFactory';
+import {YTThunkAction} from '.';
+
+export function loadAllowedExperimentalPages(): YTThunkAction {
+    return (dispatch, getState) => {
+        const login = getCurrentUserName(getState());
+        return UIFactory.getAllowedExperimentalPages(login).then((allowedExperimentalPages) => {
+            dispatch({type: GLOBAL_PARTIAL, data: {allowedExperimentalPages}});
+        });
+    };
+}
